resetPersonalBests and index.js cleaning
